+++ src/main/java/com/nisovin/shopkeepers/villagers/BlockZombieVillagerCuringListener.java
@@ -10,10 +10,10 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityTransformEvent;
 import org.bukkit.event.entity.EntityTransformEvent.TransformReason;
 import org.bukkit.event.player.PlayerInteractEntityEvent;
-import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 
 import com.nisovin.shopkeepers.Settings;
+import com.nisovin.shopkeepers.util.ItemUtils;
 import com.nisovin.shopkeepers.util.Log;
 import com.nisovin.shopkeepers.util.TextUtils;
 
@@ -28,14 +28,7 @@ public class BlockZombieVillagerCuringListener implements Listener {
 	void onZombieVillagerCureStarted(PlayerInteractEntityEvent event) {
 		if (!(event.getRightClicked() instanceof ZombieVillager)) return;
 		Player player = event.getPlayer();
-		ItemStack itemInHand;
-		if (event.getHand() == EquipmentSlot.HAND) {
-			itemInHand = player.getInventory().getItemInMainHand();
-		} else if (event.getHand() == EquipmentSlot.OFF_HAND) {
-			itemInHand = player.getInventory().getItemInOffHand();
-		} else {
-			return; // unknown
-		}
+		ItemStack itemInHand = ItemUtils.getItem(player.getInventory(), event.getHand());
 		if (itemInHand != null && itemInHand.getType() == Material.GOLDEN_APPLE) {
 			// prevent curing:
 			Log.debug(() -> "Preventing zombie villager curing at " + TextUtils.getLocationString(player.getLocation()));